requirements.txt
fastapi
uvicorn
tensorflow  # atau scikit-learn, tergantung model yang digunakan
pillow  # untuk memproses gambar
numpy
python-multipart